+++ excel_flutter_example/lib/main.dart
@@ -125,6 +125,12 @@ class _MyHomePageState extends State<MyHomePage> {
         case SelectedSection.multiFreezePanes:
           resultStatus = await ExcelGenerator.generateMultiFreezePanes();
           break;
+        case SelectedSection.hiddenColumns:
+          resultStatus = await ExcelGenerator.generateHiddenColumns();
+          break;
+        case SelectedSection.multiPageCharts:
+          resultStatus = await ExcelGenerator.generateMultiPageCharts();
+          break;
         case SelectedSection.allCharts:
           resultStatus = await ExcelGenerator.generateAllCharts();
           break;