+++ excel_flutter_example/lib/services/excel_generator.dart
@@ -6,6 +6,8 @@ import 'helpers/styles_helper.dart';
 import 'helpers/sheets_helper.dart';
 import 'helpers/protection_helper.dart';
 import 'helpers/full_demo_helper.dart';
+import 'helpers/multi_page_charts_helper.dart';
+import 'helpers/hidden_columns_helper.dart';
 
 class ExcelGenerator {
   static Future<String> generateExcelWithImage() =>
@@ -38,4 +40,10 @@ class ExcelGenerator {
 
   static Future<String> generateMultiFreezePanes() =>
       generateMultiFreezePanesHelper();
+
+  static Future<String> generateHiddenColumns() =>
+      generateHiddenColumnsHelper();
+
+  static Future<String> generateMultiPageCharts() =>
+      generateMultiPageChartsHelper();
 }